[CSM Portal] address CodeRabbit review: extract shared date resolver, fix case-type default reset bug, fix tag Enter-key removal bug
- Extract resolveRelativeDatePlaceholder out of the csm-dashboard feature
  into src/utils/, since useGetCsmCases (csm-cases) also depends on it --
  each feature's utils/ should own only its own code.
- CsmIssuesView: defaultCaseTypes now writes into the URL at most once,
  via an effect gated by a ref, instead of re-deriving in an in-memory
  useMemo override on every render. The old approach couldn't tell "fresh
  visit" apart from "user cleared the type filter back to every type" --
  both looked like "no types param" -- so clearing/Reset silently
  snapped back to the single-type default instead of staying broadened.
- TagsMultiSelect: MUI reports the same "removeOption" onChange reason
  for Enter on an already-selected, keyboard-highlighted option as it
  does for an actual chip removal (Backspace/clear). The component
  treated both as a full removal, so pressing Enter on an included tag
  silently dropped it instead of cycling it to excluded, the way a mouse
  click on that same row already does. Now distinguishes by the
  triggering key.
